+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Changelog
 ### next
+  * tpm2_loadexternal: name output to file and stdout. Changes YAML stdout output.
   * tpm2_loadexternal: ECC Public and Private PEM support.
   * tpm2_loadexternal: AES Public and Private "raw file" support.
   * tpm2_loadexternal: RSA Public and Private PEM support.